make tcpdumps optional

+++ AF_XDP-filter/netperf-namespace/runns1.sh
@@ -5,13 +5,19 @@
 ip link set lo up
 ip link set vpeer1 up
 ip addr add 10.10.0.10/16 dev vpeer1
-tcpdump -i vpeer1 -w vpeer1.tcpdump &
-tcpdump_vpeer1_pid=$!
+if [[ ! -z "${TCPDUMP}" ]]
+then
+  tcpdump -i vpeer1 -w vpeer1.tcpdump &
+  tcpdump_vpeer1_pid=$!
+fi
 sleep 6
 if [[ -z "ONLY_CRR" ]]
 then
   netperf -4 -t TCP_RR -H 10.10.0.20 -p 50000 -- -D
 fi
 netperf -4 -t TCP_CRR -H 10.10.0.20 -p 50000 -- -D
-kill -INT ${tcpdump_vpeer1_pid}
+if [[ ! -z "${TCPDUMP}" ]]
+then
+  kill -INT ${tcpdump_vpeer1_pid}
+fi
 wait
